add dsn_address_use_ip_as_name and dsn_address_local() in service_api_c.h
