+++ src/core/core/address.cpp
@@ -34,6 +34,8 @@
 # else
 # include <sys/socket.h>
 # include <netdb.h>
+# include <ifaddrs.h>
+# include <netinet/in.h>
 # include <arpa/inet.h>
 
 # if defined(__FreeBSD__)
@@ -46,6 +48,8 @@
 # include <unordered_map>
 # include <dsn/internal/singleton.h>
 # include <dsn/internal/synchronize.h>
+
+DSN_API int dsn_address_use_ip_as_name = 0;
 
 static ::dsn::utils::rw_lock_nr s_name_cache_ipv4_lock;
 static std::unordered_map<uint32_t, std::string> s_name_cache_ipv4;
@@ -95,6 +99,23 @@ DSN_API void dsn_host_to_name(const dsn_address_t* addr, /*out*/ char* name_buff
     {
     case HOST_TYPE_IPV4:
     {
+        uint32_t nip = htonl(addr->ip);
+        if (dsn_address_use_ip_as_name)
+        {
+    # if defined(_WIN32)
+            sprintf_s(
+    # else
+            std::snprintf(
+    # endif
+                name_buffer, (size_t)length,
+                "%u.%u.%u.%u",
+                nip & 0xff,
+                (nip >> 8) & 0xff,
+                (nip >> 16) & 0xff,
+                (nip >> 24) & 0xff
+                );
+        }
+
         // query cache
         s_name_cache_ipv4_lock.lock_read();
         auto it = s_name_cache_ipv4.find(addr->ip);
@@ -107,7 +128,6 @@ DSN_API void dsn_host_to_name(const dsn_address_t* addr, /*out*/ char* name_buff
         s_name_cache_ipv4_lock.unlock_read();
 
         // if not cache hit
-        uint32_t nip = htonl(addr->ip);
         // TODO: using global cache
         auto host = gethostbyaddr((char*)&nip, 4, AF_INET);
         if (host == nullptr)
@@ -202,40 +222,31 @@ DSN_API void dsn_host_from_name(dsn_host_type_t type, const char* name, /*out*/ 
     }
 }
 
-namespace dsn {
-bool rpc_address::use_ip_as_name = false;
-void rpc_address::ip_to_name(const dsn_address_t* addr, char* name_buffer, int length)
+DSN_API void dsn_address_local(/*out*/ dsn_address_t* addr, const char* network_interface)
 {
-    switch (addr->type)
-    {
-    case HOST_TYPE_IPV4:
-    {
-        uint32_t nip = htonl(addr->ip);
-# if defined(_WIN32)
-        sprintf_s(
+# ifdef _WIN32
+    dassert(false, "not implemented");
 # else
-        std::snprintf(
-# endif
-            name_buffer, (size_t)length,
-            "%u.%u.%u.%u",
-            nip & 0xff,
-            (nip >> 8) & 0xff,
-            (nip >> 16) & 0xff,
-            (nip >> 24) & 0xff
-            );
+    struct ifaddrs* ifa = nullptr;
+    getifaddrs(&ifa);
+
+    struct ifaddrs* i = ifa;
+    while (i != nullptr)
+    {
+        if (i->ifa_addr->sa_family == AF_INET && strcmp(i->ifa_name, network_interface) == 0)
+        {
+            addr->type = HOST_TYPE_IPV4;
+            addr->ip = (uint32_t)ntohl(((struct sockaddr_in *)i->ifa_addr)->sin_addr.s_addr);
+            break;
+        }
+        i = i->ifa_next;
     }
-    break;
-    case HOST_TYPE_IPV6:
-        dassert(false, "to be implemented");
-        break;
-    case HOST_TYPE_URI:
-        dassert(false, "should not be here");
-        break;
-    case HOST_TYPE_INVALID:
-        strncpy(name_buffer, "invalid_host", std::min(length, 13));
-        break;
-    default:
-        break;
+    dassert(i != nullptr, "get local ip failed, network_interface=", network_interface);
+
+    if (ifa != nullptr)
+    {
+        // remember to free it
+        freeifaddrs(ifa);
     }
-}
+#endif
 }
