+++ src/core/core/service_api_c.cpp
@@ -891,8 +891,9 @@ bool run(const char* config_file, const char* config_arguments, bool sleep_after
     }
 
     // init rpc_address::use_ip_as_name
-    dsn::rpc_address::use_ip_as_name = dsn_all.config->get_value<bool>(
+    bool use_ip_as_name = dsn_all.config->get_value<bool>(
         "network", "use_ip_as_name", false, "use ip address as rpc_address's name");
+    dsn_address_use_ip_as_name = use_ip_as_name ? 1 : 0;
     
     // setup coredump
 	auto& coredump_dir = spec.coredump_dir;
